DST-501 - Non-production warning banner for lower environments
And moving around the frontend tests
