+++ tests/test_frontend/test_report_a_suspected_breach/test_about_the_person_or_business_you_are_reporting/test_business_or_person_details.py
@@ -2,7 +2,7 @@ import re
 
 from playwright.sync_api import expect
 
-from .. import conftest, data
+from tests.test_frontend import conftest, data
 
 
 class TestAboutThePersonOrBusinessUKAddress(conftest.PlaywrightTestBase):
