+++ tests/test_frontend/test_report_a_suspected_breach/test_about_the_person_or_business_you_are_reporting/test_companies_house.py
@@ -2,7 +2,7 @@ import re
 
 from playwright.sync_api import expect
 
-from .. import conftest
+from tests.test_frontend import conftest
 
 
 class TestAreYouReportingABusinessOnCompaniesHouse(conftest.PlaywrightTestBase):
