+++ tests/test_frontend/test_report_a_suspected_breach/test_overview_of_breach/test_what_were_the_goods.py
@@ -2,7 +2,7 @@ import re
 
 from playwright.sync_api import expect
 
-from .. import conftest, data
+from tests.test_frontend import conftest, data
 
 
 class TestWhatWereTheGoods(conftest.PlaywrightTestBase):
