+++ tests/test_frontend/test_report_a_suspected_breach/test_overview_of_breach/test_when_did_you_first_suspect.py
@@ -2,7 +2,7 @@ import re
 
 from playwright.sync_api import expect
 
-from .. import conftest
+from tests.test_frontend import conftest
 
 
 class TestWhenDidYouFirstSuspect(conftest.PlaywrightTestBase):
