+++ tests/test_frontend/test_report_a_suspected_breach/test_sanctions_breach_details/test_tell_us_about_the_suspected_breach.py
@@ -2,7 +2,7 @@ import re
 
 from playwright.sync_api import expect
 
-from .. import conftest
+from tests.test_frontend import conftest
 
 
 class TestTellUsAboutTheSuspectedBreach(conftest.PlaywrightTestBase):
