+++ tests/test_frontend/test_report_a_suspected_breach/test_sanctions_breach_details/test_upload_documents.py
@@ -3,7 +3,7 @@ import re
 import pytest
 from playwright.sync_api import expect
 
-from .. import conftest, data
+from tests.test_frontend import conftest, data
 
 
 class TestUploadDocuments(conftest.PlaywrightTestBase):
