+++ tests/test_frontend/test_report_a_suspected_breach/test_supply_chain/test_about_the_supplier.py
@@ -2,7 +2,7 @@ import re
 
 from playwright.sync_api import expect
 
-from .. import conftest, data
+from tests.test_frontend import conftest, data
 
 
 class TestAboutTheSupplierUKMadeAvailableAddress(conftest.PlaywrightTestBase):
