+++ tests/test_frontend/test_report_a_suspected_breach/test_supply_chain/test_were_there_any_other_addresses_in_supply_chain.py
@@ -2,7 +2,7 @@ import re
 
 from playwright.sync_api import expect
 
-from .. import conftest
+from tests.test_frontend import conftest
 
 
 class TestWereThereAnyOtherAddressesInTheSupplyChain(conftest.PlaywrightTestBase):
