+++ tests/test_frontend/test_report_a_suspected_breach/test_your_details/test_email.py
@@ -2,7 +2,7 @@ import re
 
 from playwright.sync_api import expect
 
-from .. import conftest
+from tests.test_frontend import conftest
 
 
 class TestEmail(conftest.PlaywrightTestBase):
