+++ tests/test_frontend/test_report_a_suspected_breach/test_your_details/test_request_verify_code.py
@@ -2,7 +2,7 @@ import re
 
 from playwright.sync_api import expect
 
-from .. import conftest
+from tests.test_frontend import conftest
 
 CORRECT_CODE_DETAILS = {"email": "test@digital.gov.uk", "verify_code": "012345"}
 EMPTY_CODE_DETAILS = {"email": "test@didigtal.gov.uk", "verify_code": ""}
