+++ tests/test_unit/test_core/text_context_processors.py
@@ -1,4 +1,8 @@
-from core.context_processors import sentry_configuration_options, session_expiry_times
+from core.context_processors import (
+    environment_information,
+    sentry_configuration_options,
+    session_expiry_times,
+)
 from django.test import override_settings
 
 
@@ -25,4 +29,17 @@ def test_sentry_configuration_options(request_object):
         "SENTRY_ENVIRONMENT": "test",
         "SENTRY_ENABLE_TRACING": False,
         "SENTRY_TRACES_SAMPLE_RATE": 0.0,
+    }
+
+
+@override_settings(
+    ENVIRONMENT="test",
+    CURRENT_BRANCH="test-branch",
+    CURRENT_TAG="v1.0.0",
+)
+def test_environment_information(request_object):
+    assert environment_information(request_object) == {
+        "current_environment": "test",
+        "current_branch": "test-branch",
+        "current_tag": "v1.0.0",
     }
